Put souvenireditor files into cpp
